+++ src/channel.ts
@@ -1,4 +1,5 @@
 import { ChannelState } from './channel_state';
+import { CooldownTimer } from './CooldownTimer';
 import { MessageComposer } from './messageComposer';
 import { MessageReceiptsTracker } from './messageDelivery';
 import {
@@ -69,6 +70,7 @@ import type {
   StaticLocationPayload,
   TruncateChannelAPIResponse,
   TruncateOptions,
+  UnBanUserOptions,
   UpdateChannelAPIResponse,
   UpdateChannelOptions,
   UpdateLocationPayload,
@@ -112,6 +114,7 @@ export class Channel {
   push_preferences?: PushPreference;
   public readonly messageComposer: MessageComposer;
   public readonly messageReceiptsTracker: MessageReceiptsTracker;
+  public readonly cooldownTimer: CooldownTimer;
 
   /**
    * constructor - Create a channel
@@ -167,6 +170,8 @@ export class Channel {
         return msg && { timestampMs, msgId: msg.id };
       },
     });
+
+    this.cooldownTimer = new CooldownTimer({ channel: this });
   }
 
   /**
@@ -1580,6 +1585,7 @@ export class Channel {
         .join();
     this.data = state.channel;
     this.offlineMode = false;
+    this.cooldownTimer.refresh();
 
     if (areCapabilitiesChanged) {
       this.getClient().dispatchEvent({
@@ -1659,11 +1665,13 @@ export class Channel {
    * unbanUser - Removes the bans for a user on a channel
    *
    * @param {string} targetUserID
+   * @param {UnBanUserOptions} options
    * @returns {Promise<APIResponse>}
    */
-  async unbanUser(targetUserID: string) {
+  async unbanUser(targetUserID: string, options?: UnBanUserOptions) {
     this._checkInitialized();
     return await this.getClient().unbanUser(targetUserID, {
+      ...options,
       type: this.type,
       id: this.id,
     });
@@ -2032,6 +2040,9 @@ export class Channel {
           // 1. the message is mine
           // 2. the message is a thread reply from any user
           const preventUnreadCountUpdate = ownMessage || isThreadMessage;
+          if (ownMessage) {
+            this.cooldownTimer.refresh();
+          }
           if (preventUnreadCountUpdate) break;
 
           if (event.user?.id) {
@@ -2192,6 +2203,7 @@ export class Channel {
               event.channel?.own_capabilities ?? channel.data?.own_capabilities,
           };
           channel.data = newChannelData;
+          this.cooldownTimer.refresh();
         }
         break;
       case 'reaction.new':
@@ -2454,6 +2466,7 @@ export class Channel {
     );
 
     this.disconnected = true;
+    this.cooldownTimer.clearTimeout();
     this.state.setIsUpToDate(false);
   }
 }